+++ tuneapi/apis/__init__.py
@@ -1,7 +1,7 @@
 # Copyright © 2024-2025 Frello Technology Private Limited
 
 # model APIs
-from tuneapi.apis.model_openai import Openai, TuneModel, Groq, Mistral
+from tuneapi.apis.model_openai import Openai, TuneModel, Groq, Mistral, OpenAIProtocol
 from tuneapi.apis.model_anthropic import Anthropic
 from tuneapi.apis.model_gemini import Gemini
 from tuneapi.apis.turbo import distributed_chat, distributed_chat_async